requirements.txt

py-cord[voice]>=2.6.1
python-dotenv>=1.0.0


yt-dlp>=2025.7.21
PyNaCl>=1.5.0
async-timeout>=5.0.1
ffmpeg-python>=0.2.0


googlesearch-python>=1.3.0
beautifulsoup4>=4.13.4
requests>=2.32.4
lxml>=5.4.1


aiohttp>=3.12.15
typing-extensions>=4.14.1


cffi>=1.17.1
pycparser>=2.22


colorama>=0.4.6


json5>=0.9.14


paho-mqtt>=2.1.0
python-dateutil>=2.9.0
pytz>=2025.1


six>=1.17.0
urllib3>=2.5.0
certifi>=2025.8.3
charset-normalizer>=3.4.2
idna>=3.10


attrs>=25.3.0
frozenlist>=1.7.0
multidict>=6.6.3
propcache>=0.3.2
yarl>=1.20.1
aiohappyeyeballs>=2.6.1
aiosignal>=1.4.0


soupsieve>=2.7

